feat: moved StepperCard to components-temp directory
