+++ app/component-library/components-temp/StepperCard/StepperCard.types.ts
